Hotfix the bug caused by #4
`Object.getOwnPropertyDescriptor` won't get the descriptor from the
prototype, so if we mock the key on the prototype, it will return
undefined.

Use `Object.defineProperty` instead of assignment, because the key can't
override the getter defined on the prototype. Example below.

```
var obj = {
  get a() {
    return 1;
  }
};
var obj2 = Object.create(obj);
obj2.a = 2;
console.log(obj2.a); // return 1
```